Refactor Home.js
- Update unlitis
- Clean code
